lab_2/IS_Lab2_Python/requirements.txt
PyYAML~=6.0
multipledispatch~=0.6.0
json2xml~=3.16.0